Tidy up and re-organise components for landing page
