--- app/next-client-app/app/(public)/demo-1/page.tsx
@@ -1,20 +1,0 @@
-import Hero1 from "./hero-1";
-import Hero2 from "./hero-2";
-import BentoFeatures2 from "./features-2";
-import BentoProjects2 from "./projects-2";
-import Funders2 from "./funders-2";
-import CallToAction2 from "./CTA-2";
-import BentoFeatures from "../demo-3/bentoFeatures";
-
-export default function Demo1() {
-  return (
-    <>
-      <Hero1 />
-      <Hero2 />
-      <BentoFeatures />
-      <BentoProjects2 />
-      <Funders2 />
-      <CallToAction2 />
-    </>
-  );
-}
