--- app/next-client-app/app/(public)/demo-2/page.tsx
@@ -1,17 +1,0 @@
-import BentoProjects from "../demo-3/bentoProjects";
-import CallToAction from "./CTA";
-import Features from "./features";
-import Funders from "./funders";
-import Hero from "./hero";
-
-export default function Demo2() {
-  return (
-    <div className="space-y-12 lg:space-y-28">
-      <Hero />
-      <Features />
-      <BentoProjects />
-      <Funders />
-      <CallToAction />
-    </div>
-  );
-}
