+++ app/next-client-app/components/homepage/bentoProjects.tsx
