+++ app/next-client-app/components/homepage/features.tsx
