+++ app/next-client-app/components/homepage/funders.tsx
