+++ app/next-client-app/components/homepage/hero.tsx
